Use ProcessHandle in ProcessFixture for Java9+

+++ testing/internal-integ-testing/src/main/groovy/org/gradle/integtests/fixtures/ProcessFixture.groovy
@@ -16,11 +16,12 @@
 
 package org.gradle.integtests.fixtures
 
+
+import org.gradle.api.JavaVersion
 import org.gradle.api.internal.file.TestFiles
 import org.gradle.internal.os.OperatingSystem
 import org.gradle.process.internal.streams.SafeStreams
 import org.gradle.test.fixtures.ConcurrentTestUtil
-
 
 class ProcessFixture {
     final Long pid;
@@ -30,29 +31,74 @@ class ProcessFixture {
     }
 
     /**
-     * Forcefully kills this daemon.
+     * Forcefully kills this process.
      */
     void kill(boolean killTree) {
         println "Killing process with pid: $pid"
         if (pid == null) {
             throw new RuntimeException("Unable to force kill the process because provided pid is null!")
         }
+
+        if (JavaVersion.current().isJava9Compatible()) {
+            killJava9(killTree)
+            return
+        }
+
         if (!(OperatingSystem.current().unix || OperatingSystem.current().windows)) {
             throw new RuntimeException("This implementation does not know how to forcefully kill a process on os: " + OperatingSystem.current())
         }
         execute(killArgs(pid, killTree), killScript(pid, killTree))
     }
 
-    // Only supported on *nix platforms
+    private void killJava9(boolean killTree) {
+        def processHandle = getProcessHandle()
+        if (processHandle) {
+            if (killTree) {
+                processHandle.descendants().forEach { it.destroyForcibly() }
+            }
+            processHandle.destroyForcibly()
+        }
+    }
+
     String[] getChildProcesses() {
         if (pid == null) {
             throw new RuntimeException("Unable to get child processes because provided pid is null!")
         }
+
+        if (JavaVersion.current().isJava9Compatible()) {
+            return getChildrenJava9()
+        }
+
         if (!(OperatingSystem.current().unix)) {
-            throw new RuntimeException("This implementation does not know how to get child processes on os: " + OperatingSystem.current())
+            throw new RuntimeException("This implementation does not know how to get child processes on os: " + OperatingSystem.current() + " and Java version: " + JavaVersion.current())
         }
         String result = bash("ps -o pid,ppid -ax | awk '{ if ( \$2 == ${pid} ) { print \$1 }}'").trim()
         return result == "" ? [] : result.split("\\n")
+    }
+
+    private String[] getChildrenJava9() {
+        def processHandle = getProcessHandle()
+        if (processHandle) {
+            return processHandle.children()
+                .map { it.pid().toString() }
+                .toList()
+                .toArray(new String[0])
+        }
+        return new String[0]
+    }
+
+    /**
+     * Works only on Java 9 or later.
+     */
+    String[] getDescendants() {
+        def processHandle = getProcessHandle()
+        if (processHandle) {
+            return processHandle.descendants()
+                .map { it.pid().toString() }
+                .toList()
+                .toArray(new String[0])
+        }
+        return new String[0]
     }
 
     // Only supported on *nix platforms
@@ -61,9 +107,20 @@ class ProcessFixture {
             throw new RuntimeException("Unable to get process info because provided pids are null or empty!")
         }
         if (!(OperatingSystem.current().unix)) {
-            throw new RuntimeException("This implementation does not know how to get process info on os: " + OperatingSystem.current())
+            throw new RuntimeException("This implementation does not know how to get process info on os: " + OperatingSystem.current() + " and Java version: " + JavaVersion.current())
         }
         return bash("ps -o pid,ppid,args -p ${pids.join(' -p ')}").split("\\n")
+    }
+
+    boolean isAlive() {
+        if (pid == null) {
+            return false
+        }
+        if (JavaVersion.current().isJava9Compatible()) {
+            def processHandle = getProcessHandle()
+            return processHandle ? processHandle.isAlive() : false
+        }
+        throw new RuntimeException("ProcessFixture.isAlive() is only supported on Java 9 or later.")
     }
 
     /**
@@ -121,6 +178,17 @@ class ProcessFixture {
         } else {
             throw new IllegalStateException()
         }
+    }
+
+    private def getProcessHandle() {
+        if (!JavaVersion.current().isJava9Compatible()) {
+            throw new RuntimeException("Java 9 or later is required to get process handle.")
+        }
+        if (pid == null) {
+            throw new RuntimeException("Unable to get process handle because provided pid is null!")
+        }
+        def processHandleOptional = Class.forName("java.lang.ProcessHandle").getMethod("of", long.class).invoke(null, pid)
+        return processHandleOptional.orElse(null)
     }
 
     private static InputStream killScript(Long pid, boolean killTree) {
